RDRP-909: created freezing folder
